ButtonClickResponder moved to where it is used
