+++ web-ui/src/main/java/org/archcnl/ui/common/dialogs/OkCancelDialog.java
@@ -7,7 +7,6 @@ import com.vaadin.flow.component.dialog.Dialog;
 import com.vaadin.flow.component.orderedlayout.FlexComponent;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
-import org.archcnl.ui.common.ButtonClickResponder;
 
 public class OkCancelDialog extends Dialog {
